Add Spree::UserMethods
